pyproject.toml
[tool.poetry]
name = "hcubed"
version = "1.0.0"
description = "A random personal bot"
authors = ["vcokltfre <vcokltfre@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
python-dotenv = "^0.15.0"
loguru = "^0.5.3"
jishaku = "^1.20.0"
asyncpg = "^0.22.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
GitPython = "^3.1.14"
parsedatetime = "^2.6"

[tool.poetry.dev-dependencies]
pre-commit = "^2.11.1"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
